Refactored slightly to add an extra module containing leveldb specific classes.
