+++ modules/core/src/test/scala/io/strongtyped/funcqrs/FutureTry.scala
@@ -1,13 +1,11 @@
 package io.strongtyped.funcqrs
 
-import org.scalatest.concurrent.ScalaFutures
-
 import scala.concurrent.duration._
 import scala.concurrent.{Await, Future}
 import scala.language.postfixOps
 import scala.util.Try
 
-trait FutureTry extends ScalaFutures {
+trait FutureTry {
 
   implicit class FutureTryOps[A](fut: Future[A]) {
     def asTry = {
